refactor(sound): revert audio samples back to 44.1kHz sample rate
Previously the samples were compressed to 32kHz as it produced the same
sound quality with smaller file size.
Most of the effects in the build are 44.1kHz though and some of the
systems rely on sample rate to be exactly at that value, for example
varying sound, see /proc/get_rand_frequency().