refactor(code): replace obj-based storage UI with images and overlays
This change only affects space-based storage systems, e.g. backpacks,
pouches, bags etc.

Storage UI of these items relied on multiple objects and fixed math, as
well as used obscure naming of variables with very little documentation.
That system was replaced with new one: a single screen object represents
the storage itself, the `item_space`. It corresponds to the main
background box that represents the total space inside a storage.

To this item space two decorative "caps" are attached as overlays. They
cover open ends of the scaled item space sprite which makes item space
look like a complete container rectangle.
After that using the same horizontal scaling of the main sprite and two
caps at both ends a background is created for each individual stored
item. These backgrounds are appended as underlays to match the look of
old system.

Most of the numbers were extracted to variables and math was simplified
in order to make storage space code more readable and maintainable.

In-code documentation was added to the most complex parts of the system.