fix(code): rename `caller` variable that collides with the built-in
A new `caller` can `callee` built-in variables were added to all procs
and collide with the custom variables of the same type.
The callHook proc now uses a different variable name for the hook datum
removing the collision.

This commit is part of the migration from BYOND 515 to BYOND 516, but
does not introduce breaking changes.