feat(code): add "Enable Browser Devtools" verb for admins
With the switch to WebView2 it's now possible to initiate Chromium
devtools for any browser-based UI as long as it's enabled in the
`browser-options` for the user.

A new "Enable Browser Devtools" verb is now accessible to all admins in
the Debug tab and enables devtools access on use. After enabling,
right-click inside a browser UI (simple browser output, NanoUI, or TGUI)
and select "Inspect".