feat(code): add handling of DPI scaling for NanoUI
With BYOND's switch of the internal browser to WebView2, interfaces now
scale according to the system's DPI "Scaling" parameter.

However, since NanoUI window dimensions are passed as fixed pixel
values, scaled UIs were overflowing their respective windows.

This also introduced a change to the way UIs look compared to BYOND 515.
In versions prior to 516 IE-based rendering was not applying the system
scaling to any elements. After the update all UIs will suddenly be
larger for users with high-resolution displays if they use system
scaling, which might be an unexpected shift.

- Scale dimensions of NanoUI windows based on the user's current DPI
  scaling value by default.
- Use "Scale Interfaces for High-Resolution Displays" preference to
  determine if user wants scaling, the preference is enabled by default.
- If preference is disabled, NanoUI will be zoomed out by the inverted
  value of user's current DPI scaling. This counters the DPI scaling and
  makes UIs fully ignore it, matching pre-516 behavior that users might
  be used to.