feat(tgui): add support for Blink browser engine
As byond switched to WebView2 in BYOND 516, the internal engine for UIs
is no longer Trident, but a Chromium-based Blink.

TGUI can now properly detect both engines and will proceed with
rendering if either of the two is detected in the user agent string.

This change only affects the static TGUI html file and doesn't require
rebuilding TGUI bundles.