fix(tgui): DOM content not loading after the "ready" event
Due to the switch to Chromium-based WebView2 in BYOND 516, the default
method of making calls to BYOND via setting `window.location` introduced
an unexpected breaking change and caused the DOM contents of TGUI-based
UIs to not load after the main script in <head> was executed.

This is because modern browser engines stop loading DOM of the page if
they detect a change to `window.location`. In case of TGUI the whole
<body> of the page wasn't loading and a blank page was displayed to the
user.

The "ready" event of TGUI is now postponed until after the base html of
the page finishes loading and a `DOMContentLoaded` event is fired by the
browser engine. Changes to `window.location` are safe after that point.

Acknowledgements:
- ike709 (https://github.com/ike709) for initial issue report for
  OpenDream (https://github.com/OpenDreamProject/OpenDream/issues/1130)
  and fix request with details in Coderbus (https://discord.gg/Vh8TJp9)
  discord message (https://discord.com/channels/484170914754330625/484170915253321734/1166472904670728355)
- Cyberboss (https://github.com/Cyberboss) for the initial PR to tgstation
  addressing the issue: https://github.com/tgstation/tgstation/pull/79244
- wixoa (https://github.com/wixoaGit) for additional details on the
  cause of the issue
  (https://github.com/tgstation/tgstation/pull/79244#issuecomment-1780186812)
- itsmeow (https://github.com/itsmeow) for the event-based
  implementation of the same fix
  (https://github.com/BeeStation/BeeStation-Hornet/pull/12066)