feat(tgui): add support for DPI interface scaling based on user pref
This is only relevant for users who have high-resolution displays and
use Windows "Scaling" setting.

With BYOND's switch of the internal browser to WebView2, interfaces now
scale according to the system's DPI "Scaling" setting.

TGUI interfaces already support scaling based on `devicePixelRatio`
property, so this commit adds "backwards compatibility" for those
who are used to the way things looked without scaling.

As it is done for browser-based and NanoUI interfaces, if user sets
their "Scale Interfaces for High-Resolution Displays" preference to
"No", the interface will apply css zoom out to counter the system
scaling and mimic the old behavior.

Setting the preference to "Yes" allows the UI to scale up normally
according to the user's system setting.

TGUI bundles were rebuilt in order to apply changes.

This commit is based on the PR:
https://github.com/cmss13-devs/cmss13/pull/8723
Original author: harry (https://github.com/hry-gh)