fix: photos, images and icons in interfaces looking blurry
Starting with version 516 BYOND no longer uses IE-based browser for
rendering interfaces, so the old `-ms-interpolation-mode` css property
now doesn't apply, but is kept for backwards compatibility for clients
that still use BYOND 515.

Instead, the modern `image-rendering` property is now used. Is is
recognized by the Chromium-based WebView2 that BYOND now uses and
applies the same pixel-perfect "nearest-neighbor" algorithm for scaling
images and icons.

Fixed blurry images include:
- Photos
- Vending machine product icons
- Art library previews
- Preview image in Character Setup
- Traitor item icons in roundend screen

This commit includes a rebuild of TGUI bundles to apply the changes.