chore: remove obsolete IEChooser launch configuration
Since BYOND moved to WebView2, it is now possible to open devtools for
any UI with a simple right click as long as it's permitted by the
server. Use of IEChooser to debug browser interfaces in no longer
required.

An admin-only "Enable Browser Devtools" verb can be used to enable
access to devtools for the client, after which any interface can be
inspected via right-click -> "Inspect".