style(code): add missing quote marks around href attributes
Update all href attributes that weren't wrapped in quotes to a unified
single-quoted syntax.

This does not introduce any functional changes or fixes as non-quoted
syntax was properly recognized and automatically fixed by the browser
engine at runtime.

The change is intended to improve readability and consistency across the
codebase.