refactor(code): add byond:// scheme to corner case URLs
Some additional topic URLs and hrefs that weren't covered by the
mass-replacements are now prefixed with the `byond://` scheme.

This commit is part of a global refactoring of topic URLs into the new
format. See adjacent commits for the rest of replacements.

Some URLs couldn't be covered by the bulk replacements due to missing
src, non-standard placement, or being located in non-.dm files. These
were handled manually.
The following cases were covered:
- Ckey linkification in
  - `/proc/key_name()`
  - `/datum/client_lite/proc/key_name()`
- Old format topic links without `src` key, replaced with modern format
  `href='?\ref[src];` -> `href='byond://?src=\ref[src];`
  - 3 links in `update()` proc of datum/hSB
  - 5 links in `/obj/item/pen/crayon/proc/update_popup()`
- Document overwrite scripts in:
  - file `code/js/wiki_html.js`
  - file `code/js/wiki_text.js`
- Admin quick jump links in:
  - `/mob/extra_admin_link()` (EYE links)
  - `/atom/movable/extra_admin_link()` (JMP links)
  - `/mob/observer/ghost/extra_admin_link()` (BDY links)
- HTML form's `action` in `/datum/watchlist/proc/Show()`
- Window location hrefs in inline scripts in:
  - `/proc/send_asset()`
  - `/proc/send_asset_list()`
- Link replacement regex in `/obj/item/paper`'s `field_link_regex`
- NanoUI's main `generateHref` function in `nano/js/nano_utility.js`
- HTML form's `action` in file `nano/templates/accounts_terminal.tmpl`
- An example href in file `tgui/docs/tgui-for-custom-html-popups.md`

Details:
In BYOND, all URLs used for communication between browser controls and
the server should include the `byond://` scheme. For example:
  `?src=[0x...]&action=1`
becomes:
  `byond://src=[0x...]&action=1`

This is because BYOND relies on the `byond://` scheme to be present for
correct link interpretation in browser controls.
The short `?topic` form is technically supported via a fragile
workaround that interprets it as byond://, but this behavior is not
guaranteed, especially since it briefly broke in BYOND 516.

Using `byond://` avoids relying on legacy compatibility hacks and
ensures the link is processed as a topic call instead of a relative
document navigation URL.

References:
- Forum thread: https://www.byond.com/forum/post/2958145#comment26841953
- Clarification from Lummox JR (BYOND dev) in BYONDiscord: https://discord.com/channels/725444629172060262/725458598213451879/1361436640064569405
  - BYONDiscord invite: https://discord.com/invite/WUetUEVdk4