fix(tgui): dark theme tgchat flickering white on new messages
In BYOND 516 tg-chat started visibly flickering to white background
when new messages were received by the client.

This was due to BYOND momentarily changing is-visible/is-disabled
properties of the output container, which displayed the white background
for a split second causing user-visible flicker.

This was resolved by changing the approach to swapping out the panes for
native BYOND output and tg-chat output.
A `CHILD` container is now used and displays only one chat at a time,
while the other one stays unmapped from the user UI, and so completely
invisible.

TGUI bundles were rebuilt in order to apply changes on JS side.

Acknowledgements:
- ShadowLarkens (https://github.com/ShadowLarkens) for the original
  research of the issue and implementation of the fix.