fix(code): ping command not working in long-running rounds
In order to estimate network delay, `Ping` command uses a trick of
letting the server invoke another hidden command on behalf of the user.

The estimation is done by simple subtraction: timestamp that is passed
to the server by the `Ping` command is subtracted from the actual server
time.

Timestamp was simply added to the command by embedding a number in a
string, which uses default num2text conversion rules and switches to
scientific notation (e.g. 1.23e+09) when number goes beyond six digits.

Scientific notation numbers can't be passed as an `as num` parameter,
and so the verb was returning an error. The error was printed in classic
chat and "Options and Messages" window of the client, but not visible
from the TGUI chat.
This meant that `Ping` seemingly stopped working if round has been on
standby for 27 or more hours and someone tried using it after that.

The command now uses explicit `num2text` with 10 significant digits.
This allows for numbers up to 10 billion to be passed without converting
or losing precision.
Assuming ideal calculation and given that these are deciseconds, should
be sufficient for a 30 year round.