feat(code): item worths list cleanup and adjustments
Sorted things and now they are a bit easier to find:
- Weapon list sorted: Added "sorting" for pulse, accelerator, magic,
taser, energy, sniper weapons.
- A silencer has been placed near the list of weapons, as this item is
needed solely for the holdout of the gun to which it is directly
attached.
- A small list of ammunition was also placed near the weapon. In the
future it will be possible to break them down into prices, but for now
we have what we have.
- A category was created for gas masks, armor and helmets.
- A category was created for shields, RIG modules, “things that fall
under the profession”, various tools (regular and surgical), boards,
musical instruments, magic items, canes, soaps, pinpointers, disks,
handcuffs, tanks, modules, cartridges, illegal items, hand teleporter.

Adjusted some prices based on personal experience and preferences:
- The price of the magic staff has been added. The price is small, as
any of the crew members cannot (should not) use them, which creates an
illusion on “toy”
- Prices for taser weapons, specifically prices for carbine and rifle
have been added.
- Pistol and revolver prices. Some have been tweaked and some have been
added.
- The price of a “lambda” that could be through much suffering and agony
on the part of the merchant, who without fraudulent tactics in his life
will not be able to acquire this item, which god forbid someone buys for
half the price.
- The same goes for the “infinite cell”, and while it is indeed
infinite, let's not forget the very small energy capacity within it,
which creates some limitations on usage.
- Slightly reduced the price of meson glasses.
- Slightly changed the prices of various boots.
- Changed the price of armor plate and medals.
- Added extensive prices for improved surgical instruments as well as
organs.
- Added prices for soaps.
- Added prices for different cartridges (on the PDA).
- Added prices for collectible figures of station personnel.
- Added prices for cats because it's **very important.**

PR #12